Move Pool types and move snapshot block functions
