+++ src/services/balancer/subgraph/balancer-subgraph.service.ts
@@ -1,6 +1,3 @@
-import { Network } from '@balancer-labs/sdk';
-
-import { networkId } from '@/composables/useNetwork';
 import { rpcProviderService as _rpcProviderService } from '@/services/rpc-provider/rpc-provider.service';
 
 import { balancerSubgraphClient } from './balancer-subgraph.client';
@@ -30,22 +27,6 @@ export default class BalancerSubgraphService {
     this.poolSwaps = new PoolSwaps(this);
     this.poolSnapshots = new PoolSnapshots(this);
     this.tradePairSnapshots = new TradePairSnapshots(this);
-  }
-
-  public get blockTime(): number {
-    switch (networkId.value) {
-      case Network.MAINNET:
-        return 13;
-      case Network.POLYGON:
-        return 2;
-      case Network.ARBITRUM:
-        return 3;
-      case Network.KOVAN:
-        // Should be ~4s but this causes subgraph to return with unindexed block error.
-        return 1;
-      default:
-        return 13;
-    }
   }
 }
