+++ src/services/pool/types.ts
@@ -2,13 +2,6 @@ import { BigNumber } from 'ethers';
 
 import { Address } from '@/types';
 
-export type QueryArgs = Record<string, any>;
-export type QueryAttrs = Record<string, any>;
-export type QueryBuilder = (
-  args?: QueryArgs,
-  attrs?: QueryAttrs
-) => Record<string, any>;
-
 export enum PoolType {
   Weighted = 'Weighted',
   Investment = 'Investment',
@@ -17,7 +10,6 @@ export enum PoolType {
   StablePhantom = 'StablePhantom',
   LiquidityBootstrapping = 'LiquidityBootstrapping'
 }
-export type TimeTravelPeriod = '24h';
 
 export interface PoolToken {
   address: string;
