+++ src/services/web3/web3.service.ts
@@ -129,32 +129,8 @@ export default class Web3Service {
     }
   }
 
-  public get blockTime(): number {
-    switch (networkId.value) {
-      case Network.MAINNET:
-        return 13;
-      case Network.POLYGON:
-        return 2;
-      case Network.ARBITRUM:
-        return 3;
-      case Network.KOVAN:
-        // Should be ~4s but this causes subgraph to return with unindexed block error.
-        return 1;
-      default:
-        return 13;
-    }
-  }
-
-  public async getTimeTravelBlock(period: TimeTravelPeriod): Promise<number> {
-    const currentBlock = await rpcProviderService.getBlockNumber();
-    const blocksInDay = Math.round(twentyFourHoursInSecs / this.blockTime);
-
-    switch (period) {
-      case '24h':
-        return currentBlock - blocksInDay;
-      default:
-        return currentBlock - blocksInDay;
-    }
+  public async getCurrentBlock(): Promise<number> {
+    return await rpcProviderService.getBlockNumber();
   }
 
   public async callStatic<T>(
